[VIC-17] [VIC-23] [VIC-24] [VIC-26] Macroblocks motion vectors (#24)
* Initial implementation of macroblock information handling

* Moved macroblocks to the right panel

* Fix i-frame CSS bug; add frame number input & display; add loaders to macroblock buttons

* Removed best parameters endpoint and button, decreased precision in slider to MB

* Refactor macroblock extraction to use pipeline, remove unused endpoints, improve block mapping and selection indication, and fix grid rendering bugs

* Add smooth transitions and auto-scroll to info section

* Add macroblock history with reference and moved blocks

* Update macroblock grid & motion vector rendering, improve history visuals and CSS

* Add disappear mode to hide unchecked block types

* Added VideoParameters view, fixed compression views

* fix bug related to bandwidth, updated startup to ignore gitkeeps

* fix: remove print

---------

+++ backend/startup.py
@@ -1,4 +1,4 @@
-from compressor import models  # zmień 'yourapp' na nazwę swojej aplikacji Django
+from compressor import models
 import os
 import sys
 import shutil
@@ -8,11 +8,24 @@ models.Video.objects.all().delete()
 reserved_filenames = os.listdir(os.path.join("static", "original_videos"))
 reserved_frame_dirs = set(map(lambda x: os.path.splitext(x)[0], reserved_filenames))
 
-# for file in os.listdir(os.path.join("static", "compressed_videos")):
-#     os.remove(os.path.join("static", "compressed_videos", file))
+for file in os.listdir(os.path.join("static", "compressed_videos")):
+    if file != ".gitkeep":
+        os.remove(os.path.join("static", "compressed_videos", file))
 for file in os.listdir(os.path.join("static", "frames")):
-    if file not in reserved_frame_dirs:
+    if file not in reserved_frame_dirs and file != ".gitkeep":
         shutil.rmtree(os.path.join("static", "frames", file))
+
+folder = os.path.join("static", "macroblocks")
+
+for name in os.listdir(folder):
+    path = os.path.join(folder, name)
+    if name == ".gitkeep":
+        continue
+
+    if os.path.isfile(path) or os.path.islink(path):
+        os.remove(path)
+    elif os.path.isdir(path):
+        shutil.rmtree(path)
 
 video_titles = {
     "example1.y4m": "Witcher 3",
@@ -39,12 +52,6 @@ for vid in reserved_filenames:
             height=1080,
             original_filename=vid,
             title=title,
-            preset="medium",
-            aq_mode=1,
-            aq_strength=1.0,
-            bf=2,
-            crf=25,
-            gop_size=60,
             size=size
         )
         video_dir = os.path.join("static", "frames", name)
